11/10/2018 Schedule and ScheduleTask rework
Even after adding features for Schedule on 18/08/18, the design for Schedule and ScheduleTask wasn't good hence it was very confusing to work with it, a rework was needed to simplify some things. In this update Schedule and ScheduleTask was reworked to make it more simplified and easy to work with.

Changes in the update:

- Schedule: removed reScheduleDeltaTime
- Schedule: removed nextSchedule()
- ScheduleTask: changed from abstract class to interface
- ScheduleTask: added SchduleTask.getNextDate()
- Created ScheduleCC and Removed ScheduleCreateCC and ScheduleDBDeleteCC
- Created ReScheduler
- ScheduleManager.notifyUpdate name changed to ScheduleManager.notifyDBUpdate